refactor: Divide api.impl into mutiple files as process, standalone, eventhub
