+++ adapter/internal/oasparser/model/project_structs.go
@@ -1,0 +1,128 @@
+/*
+ *  Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package model
+
+import "github.com/wso2/product-microgateway/adapter/pkg/synchronizer"
+
+// ProjectAPI contains the extracted from an API project zip
+type ProjectAPI struct {
+	APIYaml            APIYaml
+	APIEnvProps        map[string]synchronizer.APIEnvProps
+	Deployments        []Deployment
+	OpenAPIJsn         []byte
+	InterceptorCerts   []byte
+	APIType            string // read from api.yaml and formatted to upper case
+	APILifeCycleStatus string // read from api.yaml and formatted to upper case
+	OrganizationID     string // read from api.yaml or config
+
+	//UpstreamCerts cert filename -> cert bytes
+	UpstreamCerts map[string][]byte
+	//EndpointCerts url -> cert filename
+	EndpointCerts map[string]string
+}
+
+// EndpointSecurity contains parameters of endpoint security at api.json
+type EndpointSecurity struct {
+	Password         string            `json:"password,omitempty" mapstructure:"password"`
+	Type             string            `json:"type,omitempty" mapstructure:"type"`
+	Enabled          bool              `json:"enabled,omitempty" mapstructure:"enabled"`
+	Username         string            `json:"username,omitempty" mapstructure:"username"`
+	CustomParameters map[string]string `json:"customparameters,omitempty" mapstructure:"customparameters"`
+}
+
+// APIEndpointSecurity represents the structure of endpoint_security param in api.yaml
+type APIEndpointSecurity struct {
+	Production EndpointSecurity `json:"production,omitempty"`
+	Sandbox    EndpointSecurity `json:"sandbox,omitempty"`
+}
+
+// ApimMeta represents APIM meta information of files received from APIM
+type ApimMeta struct {
+	Type    string `yaml:"type" json:"type"`
+	Version string `yaml:"version" json:"version"`
+}
+
+// DeploymentEnvironments represents content of deployment_environments.yaml file
+// of an API_CTL Project
+type DeploymentEnvironments struct {
+	ApimMeta
+	Data []Deployment `yaml:"data"`
+}
+
+// Deployment represents deployment information of an API_CTL project
+type Deployment struct {
+	DisplayOnDevportal    bool   `yaml:"displayOnDevportal"`
+	DeploymentVhost       string `yaml:"deploymentVhost"`
+	DeploymentEnvironment string `yaml:"deploymentEnvironment"`
+}
+
+// EndpointCertificatesDetails represents content of endpoint_certificates.yaml file
+// of an API_CTL Project
+type EndpointCertificatesDetails struct {
+	ApimMeta
+	Data []EndpointCertificate `json:"data"`
+}
+
+// EndpointCertificate represents certificate information of an API_CTL project
+type EndpointCertificate struct {
+	Alias       string `json:"alias"`
+	Endpoint    string `json:"endpoint"`
+	Certificate string `json:"certificate"`
+}
+
+// APIYaml contains everything necessary to extract api.json/api.yaml file
+// To support both api.json and api.yaml we convert yaml to json and then use json.Unmarshal()
+// Therefore, the params are defined to support json.Unmarshal()
+type APIYaml struct {
+	ApimMeta
+	Data struct {
+		ID                         string   `json:"Id,omitempty"`
+		Name                       string   `json:"name,omitempty"`
+		Context                    string   `json:"context,omitempty"`
+		Version                    string   `json:"version,omitempty"`
+		RevisionID                 int      `json:"revisionId,omitempty"`
+		APIType                    string   `json:"type,omitempty"`
+		LifeCycleStatus            string   `json:"lifeCycleStatus,omitempty"`
+		EndpointImplementationType string   `json:"endpointImplementationType,omitempty"`
+		AuthorizationHeader        string   `json:"authorizationHeader,omitempty"`
+		SecurityScheme             []string `json:"securityScheme,omitempty"`
+		OrganizationID             string   `json:"organizationId,omitempty"`
+		EndpointConfig             struct {
+			EndpointType                 string              `json:"endpoint_type,omitempty"`
+			LoadBalanceAlgo              string              `json:"algoCombo,omitempty"`
+			LoadBalanceSessionManagement string              `json:"sessionManagement,omitempty"`
+			LoadBalanceSessionTimeOut    string              `json:"sessionTimeOut,omitempty"`
+			APIEndpointSecurity          APIEndpointSecurity `json:"endpoint_security,omitempty"`
+			RawProdEndpoints             interface{}         `json:"production_endpoints,omitempty"`
+			ProductionEndpoints          []EndpointInfo
+			ProductionFailoverEndpoints  []EndpointInfo `json:"production_failovers,omitempty"`
+			RawSandboxEndpoints          interface{}    `json:"sandbox_endpoints,omitempty"`
+			SandBoxEndpoints             []EndpointInfo
+			SandboxFailoverEndpoints     []EndpointInfo `json:"sandbox_failovers,omitempty"`
+			ImplementationStatus         string         `json:"implementation_status,omitempty"`
+		} `json:"endpointConfig,omitempty"`
+	} `json:"data"`
+}
+
+// EndpointInfo holds config values regards to the endpoint
+type EndpointInfo struct {
+	Endpoint string `json:"url,omitempty"`
+	Config   struct {
+		ActionDuration string `json:"actionDuration,omitempty"`
+		RetryTimeOut   string `json:"retryTimeOut,omitempty"`
+	} `json:"config,omitempty"`
+}
